Docs maintenance (#106)
* Update ReadTheDocs build config

Upgraded the OS to Ubuntu 24.04 and updated the Python toolchain to
mambaforge-23.11 in the .readthedocs.yaml file. This ensures
compatibility with the latest environment settings available on
readthedocs.

* Update Sphinx & optional dependencies version pins

Updated Sphinx and optional dependencies packages pinned versions in
`environment-rtd.yaml` and `SalishSeaTools/environment.yaml` in preparation for
Python 3.13 environment.

* Drop Sphinx source_suffix config

The deprecated list of strings appears to have been removed in
Sphinx=8. The default `source_suffix = {'.rst': 'restructuredtext'}` is
appropriate for this project.

* Update cross-reference link in Marlin commands docs

Changed the reference link from `PullChangesFromNEMOsvn` to
`PullChangesFromNEMO-3.6svn` in the Marlin commands documentation. This directs
users to the correct details for using the `marlin` command.

* Refactor docstrings in geo_tools.get_ij_coordinates()

Simplified and standardized the format of the docstrings for
`geo_tools.get_ij_coordinates` function to improve readability. Changed the type
 annotations to inline format for better consistency with other parts of the
 code and to resolve warnings from Sphinx autodoc extension.

* Fix many broken links